// server/src/types/attributes.d.ts
type Attributes = {
  finishing: number, crossing: number, dribbling: number, heading: number, tackling: number, marking: number, passing: number, free_kick: number,
  acceleration: number, agility: number, strength: number, jumping: number,
  vision: number, decision: number, positioning: number, antecipation: number, aggression: number,
  reflexes: number, handling: number, diving: number
}

export default Attributes;